+++ openlane/scripts/openroad/ioplacer.tcl
@@ -15,50 +15,51 @@ source $::env(SCRIPTS_DIR)/openroad/common/io.tcl
 read_current_odb
 
 if { [info exists ::env(CONTEXTUAL_IO_FLAG)] } {
-	read_lef $::env(placement_tmpfiles)/top_level.lef
+    read_lef $::env(placement_tmpfiles)/top_level.lef
 }
 
-if { [info exists ::env(FP_IO_HLENGTH)] } {
-	set_pin_length -hor_length $::env(FP_IO_HLENGTH)
+if { [info exists ::env(IO_PIN_H_LENGTH)] } {
+    set_pin_length -hor_length $::env(IO_PIN_H_LENGTH)
 }
 
-if { [info exists ::env(FP_IO_VLENGTH)] } {
-	set_pin_length -ver_length $::env(FP_IO_VLENGTH)
+if { [info exists ::env(IO_PIN_V_LENGTH)] } {
+    set_pin_length -ver_length $::env(IO_PIN_V_LENGTH)
 }
 
-if { $::env(FP_IO_HEXTEND) != "0"} {
-	set_pin_length_extension -hor_extension $::env(FP_IO_HEXTEND)
+if { $::env(IO_PIN_H_EXTENSION) != "0"} {
+    set_pin_length_extension -hor_extension $::env(IO_PIN_H_EXTENSION)
 }
 
-if { $::env(FP_IO_VEXTEND) != "0"} {
-	set_pin_length_extension -ver_extension $::env(FP_IO_VEXTEND)
+if { $::env(IO_PIN_V_EXTENSION) != "0"} {
+    set_pin_length_extension -ver_extension $::env(IO_PIN_V_EXTENSION)
 }
 
-if {$::env(FP_IO_VTHICKNESS_MULT) != "" && $::env(FP_IO_HTHICKNESS_MULT) != ""} {
-	set_pin_thick_multiplier -hor_multiplier $::env(FP_IO_HTHICKNESS_MULT) \
-		-ver_multiplier $::env(FP_IO_VTHICKNESS_MULT)
+if {$::env(IO_PIN_V_THICKNESS_MULT) != "" && $::env(IO_PIN_H_THICKNESS_MULT) != ""} {
+    set_pin_thick_multiplier\
+        -hor_multiplier $::env(IO_PIN_H_THICKNESS_MULT) \
+        -ver_multiplier $::env(IO_PIN_V_THICKNESS_MULT)
 }
 
 set arg_list [list]
-if { $::env(FP_PPL_MODE) == "random_equidistant" } {
-	lappend arg_list -random
+if { $::env(IO_PIN_PLACEMENT_MODE) == "random_equidistant" } {
+    lappend arg_list -random
 }
 
-if { [info exists ::env(FP_IO_MIN_DISTANCE)] } {
-	lappend arg_list -min_distance $::env(FP_IO_MIN_DISTANCE)
+if { [info exists ::env(IO_PIN_MIN_DISTANCE)] } {
+    lappend arg_list -min_distance $::env(IO_PIN_MIN_DISTANCE)
 }
 
-if { $::env(FP_PPL_MODE) == "annealing" } {
-	lappend arg_list -annealing
+if { $::env(IO_PIN_PLACEMENT_MODE) == "annealing" } {
+    lappend arg_list -annealing
 }
 
 set HMETAL $::env(FP_IO_HLAYER)
 set VMETAL $::env(FP_IO_VLAYER)
 
 log_cmd place_pins {*}$arg_list \
-	-random_seed 42 \
-	-hor_layers $HMETAL \
-	-ver_layers $VMETAL
+    -random_seed 42 \
+    -hor_layers $HMETAL \
+    -ver_layers $VMETAL
 
 write_views