+++ Changelog.md
@@ -28,6 +28,10 @@
   * `PDN_OBSTRUCTIONS` and `ROUTING_OBSTRUCTIONS` are now lists of tuples
     instead of variable-length Tcl-style lists (AKA: strings).
 
+* `Odb.CustomIOPlacement`
+
+  * All variables prefixed `FP_IO_` have been renamed, now prefixed `IO_PIN_`.
+
 * `OpenROAD.*`
 
   * Added `PNR_CORNERS`. An override for `DEFAULT_CORNER` for PnR steps except
@@ -75,9 +79,24 @@
   * Creates three reports to help verify that the RC values used for estimation
     are set correctly.
 
+* `OpenROAD.Floorplan`
+
+  * Added `FP_FLIP_SITES`: allows sites in floorplans to be flipped. Useful in
+    niche alignment scenarios where single-height cells have ground at the south
+    side and double-height cells have power at the south side, causing a short.
+    In that situation, flipping the sites for single-height cells resolves the
+    issue.
+
+* `OpenROAD.GeneratePDN`
+
+  * All variables prefixed `FP_PDN_` have been renamed to be prefixed simply
+    `PDN`. Backwards compatibility wrapper code has been added for `PDN_CFG`
+    files.
+
 * `OpenROAD.GlobalPlacement`
 
   * Added optional variable `PL_ROUTABILITY_MAX_DENSITY_PCT`
+
   * Added optional variable `PL_KEEP_RESIZE_BELOW_OVERFLOW`
 
   * Corrected `GPL_CELL_PADDING` to be an integer.